Start splitting up ride.h
