+++ src/openrct2/core/DataSerialiserTraits.h
@@ -16,7 +16,7 @@
 #include "../network/network.h"
 #include "../object/Object.h"
 #include "../peep/Guest.h"
-#include "../ride/Ride.h"
+#include "../ride/RideColour.h"
 #include "../ride/TrackDesign.h"
 #include "../world/Location.hpp"
 #include "../world/TileElement.h"
