+++ src/openrct2/libopenrct2.vcxproj
@@ -383,7 +383,9 @@
     <ClInclude Include="ride\gentle\meta\SpiralSlide.h" />
     <ClInclude Include="ride\Ride.h" />
     <ClInclude Include="ride\RideAudio.h" />
+    <ClInclude Include="ride\RideColour.h" />
     <ClInclude Include="ride\RideData.h" />
+    <ClInclude Include="ride\RideEntry.h" />
     <ClInclude Include="ride\RideRatings.h" />
     <ClInclude Include="ride\RideTypes.h" />
     <ClInclude Include="ride\ShopItem.h" />
