+++ src/openrct2/object/ObjectList.h
@@ -9,12 +9,6 @@
 
 #pragma once
 
-#include "../ride/Ride.h"
-#include "../world/Banner.h"
-#include "../world/Entrance.h"
-#include "../world/Footpath.h"
-#include "../world/Scenery.h"
-#include "../world/Water.h"
 #include "Object.h"
 #include "ObjectLimits.h"
