+++ src/openrct2/object/ObjectRepository.h
@@ -11,7 +11,7 @@
 
 #include "../common.h"
 #include "../object/Object.h"
-#include "../ride/Ride.h"
+#include "RideObject.h"
 
 #include <memory>
 #include <vector>
