+++ src/openrct2/ride/Ride.cpp
@@ -36,6 +36,7 @@
 #include "../object/MusicObject.h"
 #include "../object/ObjectList.h"
 #include "../object/ObjectManager.h"
+#include "../object/RideObject.h"
 #include "../object/StationObject.h"
 #include "../paint/VirtualFloor.h"
 #include "../peep/Peep.h"
@@ -48,6 +49,7 @@
 #include "../windows/Intent.h"
 #include "../world/Banner.h"
 #include "../world/Climate.h"
+#include "../world/Entrance.h"
 #include "../world/Footpath.h"
 #include "../world/Location.hpp"
 #include "../world/Map.h"
@@ -57,6 +59,7 @@
 #include "CableLift.h"
 #include "RideAudio.h"
 #include "RideData.h"
+#include "RideEntry.h"
 #include "ShopItem.h"
 #include "Station.h"
 #include "Track.h"
