+++ src/openrct2/ride/RideColour.h
@@ -1,5 +1,5 @@
 /*****************************************************************************
- * Copyright (c) 2014-2020 OpenRCT2 developers
+ * Copyright (c) 2014-2021 OpenRCT2 developers
  *
  * For a complete list of all authors, please refer to contributors.md
  * Interested in contributing? Visit https://github.com/OpenRCT2/OpenRCT2
@@ -11,8 +11,16 @@
 
 #include "../common.h"
 
-struct Vehicle;
-struct Ride;
+struct TrackColour
+{
+    uint8_t main;
+    uint8_t additional;
+    uint8_t supports;
+};
 
-Vehicle* cable_lift_segment_create(
-    Ride& ride, int32_t x, int32_t y, int32_t z, int32_t direction, uint16_t var_44, int32_t remaining_distance, bool head);
+struct vehicle_colour
+{
+    uint8_t main;
+    uint8_t additional_1;
+    uint8_t additional_2;
+};
