+++ src/openrct2/ride/Station.cpp
@@ -13,6 +13,7 @@
 #include "../peep/Guest.h"
 #include "../scenario/Scenario.h"
 #include "../world/Location.hpp"
+#include "RideEntry.h"
 #include "Track.h"
 #include "Vehicle.h"
