+++ src/openrct2/ride/gentle/Dodgems.cpp
@@ -12,6 +12,7 @@
 #include "../../paint/Paint.h"
 #include "../../paint/Supports.h"
 #include "../../util/Util.h"
+#include "../Ride.h"
 #include "../Track.h"
 #include "../TrackPaint.h"
