+++ src/openrct2/ride/gentle/FerrisWheel.cpp
@@ -12,6 +12,8 @@
 #include "../../paint/Paint.h"
 #include "../../paint/Supports.h"
 #include "../../peep/Guest.h"
+#include "../Ride.h"
+#include "../RideEntry.h"
 #include "../Track.h"
 #include "../TrackPaint.h"
 #include "../Vehicle.h"
