+++ src/openrct2/ride/gentle/SpaceRings.cpp
@@ -13,6 +13,7 @@
 #include "../../paint/Paint.h"
 #include "../../paint/Supports.h"
 #include "../../peep/Guest.h"
+#include "../RideEntry.h"
 #include "../Track.h"
 #include "../TrackPaint.h"
 #include "../Vehicle.h"
