+++ src/openrct2/ride/gentle/SpiralSlide.cpp
@@ -11,6 +11,7 @@
 #include "../../object/StationObject.h"
 #include "../../paint/Paint.h"
 #include "../../paint/Supports.h"
+#include "../Ride.h"
 #include "../Track.h"
 #include "../TrackPaint.h"
