+++ src/openrct2/ride/thrill/LaunchedFreefall.cpp
@@ -12,6 +12,7 @@
 #include "../../interface/Viewport.h"
 #include "../../paint/Paint.h"
 #include "../../paint/Supports.h"
+#include "../Ride.h"
 #include "../Track.h"
 #include "../TrackPaint.h"
 #include "../Vehicle.h"
