+++ src/openrct2/ride/thrill/RotoDrop.cpp
@@ -11,6 +11,7 @@
 #include "../../interface/Viewport.h"
 #include "../../paint/Paint.h"
 #include "../../paint/Supports.h"
+#include "../Ride.h"
 #include "../Track.h"
 #include "../TrackPaint.h"
 #include "../Vehicle.h"
